+++ lib/api/response_models/app_response.dart
@@ -15,7 +15,7 @@ class AppResponse extends ApiResponse {
       : id = json["id"],
         name = json["name"],
         serviceName = json["service_name"],
-        icon = IconData(json["icon"], fontFamily: 'MaterialIcons'),
+        icon = json["icon"] != null ? IconData(json["icon"], fontFamily: 'MaterialIcons') : null,
         color = json["color"].toString().parseColor(),
         repository = json["repository"],
         repositoryBranch = json["repository_branch"];
