+++ lib/views/auth/auth_page.dart
@@ -1,11 +1,11 @@
+import 'dart:html' as html;
+
+import 'package:client_common/api/application_api.dart';
+import 'package:client_common/api/response_models/app_response.dart';
 import 'package:client_common/models/auth_model.dart';
-import 'package:client_common/models/user_application_model.dart';
-import 'package:client_common/navigator/common_navigator.dart';
-import 'package:client_common/views/loading_button.dart';
+import 'package:client_common/views/auth/auth_page_form.dart';
 import 'package:client_common/views/simple_page.dart';
-import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
 import 'package:lenra_components/component/lenra_text.dart';
 import 'package:lenra_components/lenra_components.dart';
 import 'package:logging/logging.dart';
@@ -18,315 +18,100 @@ class AppAuthPage extends StatefulWidget {
 
 class _AppAuthPageState extends State<AppAuthPage> {
   final logger = Logger('AuthPage');
-  final _formKey = GlobalKey<FormState>();
 
-  bool isRegisterPage = true;
-  String email = "";
-  String password = "";
-
-  bool keep = false;
-  bool _hidePassword = true;
   var themeData = LenraThemeData();
-  bool isLogging = false;
 
   String? redirectTo;
+  String? appServiceName;
 
   @override
   Widget build(BuildContext context) {
-    isLogging = context.select<AuthModel, bool>(
-      (m) => isRegisterPage ? m.registerStatus.isFetching() : m.loginStatus.isFetching(),
-    );
-
     redirectTo = context.read<AuthModel>().redirectToRoute;
+    print("redirectTo $redirectTo");
     RegExp regExp = RegExp(r"app\/([a-fA-F0-9-]{36})");
     final match = regExp.firstMatch(redirectTo ?? "/");
-    String? appServiceName = match?.group(1);
-
-    if (appServiceName != null) {
-      context.read<UserApplicationModel>().getAppByServiceName(appServiceName);
-      // TODO continue this code
-    }
+    appServiceName = match?.group(1);
+    print("appServiceName $appServiceName");
 
     return SimplePage(
-      header: Container(),
+      header: appServiceName != null ? Container() : null,
       child: LenraFlex(
         spacing: 32,
         direction: Axis.vertical,
         crossAxisAlignment: CrossAxisAlignment.center,
+        mainAxisAlignment: MainAxisAlignment.center,
+        fillParent: true,
         children: [
-          Container(
-            width: 100,
-            height: 100,
-            alignment: Alignment.center,
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(6),
-              color: Color.fromARGB(255, 187, 234, 255),
-            ),
-            child: LenraText(
-              text: "A",
-              style: TextStyle(
-                fontSize: 80,
-                fontWeight: FontWeight.bold,
-                color: LenraColorThemeData.lenraBlue,
-              ),
-            ),
-          ),
-          LenraFlex(
-            direction: Axis.vertical,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              LenraText(
-                text: "App Name",
-                style: themeData.lenraTextThemeData.headline1,
-              ),
-              LenraFlex(
-                spacing: 8,
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  LenraText(text: "by"),
-                  Image.asset(
-                    "assets/images/logo-horizontal-black.png",
-                    scale: 1.25,
-                  ),
-                  IconButton(
-                    onPressed: () {},
-                    icon: Icon(Icons.info_outline),
-                  ),
-                ],
-              ),
-            ],
-          ),
-          Container(
-            decoration: BoxDecoration(
-              color: LenraColorThemeData.greySuperLight,
-              borderRadius: BorderRadius.circular(32),
-            ),
-            child: LenraFlex(
-              spacing: 16,
+          appHeader(),
+          AuthPageForm(),
+        ],
+      ),
+    );
+  }
+
+  Widget appHeader() {
+    if (appServiceName != null) {
+      return FutureBuilder(
+        future: ApplicationApi.getAppByServiceName(appServiceName!),
+        builder: (context, AsyncSnapshot<AppResponse> snapshot) {
+          if (snapshot.connectionState == ConnectionState.waiting) {
+            return CircularProgressIndicator();
+          } else {
+            return LenraFlex(
+              direction: Axis.vertical,
+              crossAxisAlignment: CrossAxisAlignment.center,
+              spacing: 32,
               children: [
-                TextButton(
-                  onPressed: () {
-                    setState(() {
-                      isRegisterPage = true;
-                    });
-                  },
-                  child: LenraText(
-                    text: "Register",
-                    style: isRegisterPage
-                        ? TextStyle(color: LenraColorThemeData.lenraWhite)
-                        : TextStyle(color: LenraColorThemeData.lenraBlue),
+                Container(
+                  width: 100,
+                  height: 100,
+                  alignment: Alignment.center,
+                  decoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(6),
+                    color: Color.fromARGB(255, 187, 234, 255),
                   ),
-                  style: ButtonStyle(
-                    backgroundColor: MaterialStateProperty.resolveWith((states) {
-                      if (isRegisterPage) {
-                        return LenraColorThemeData.lenraBlue;
-                      }
-                      return LenraColorThemeData.greySuperLight;
-                    }),
-                    shape: MaterialStateProperty.all(
-                      RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(32),
-                      ),
+                  child: LenraText(
+                    text: snapshot.data?.name?[0].toUpperCase() ?? "A",
+                    style: TextStyle(
+                      fontSize: 80,
+                      fontWeight: FontWeight.bold,
+                      color: LenraColorThemeData.lenraBlue,
                     ),
-                    fixedSize: MaterialStateProperty.all(Size(136, 36)),
                   ),
                 ),
-                TextButton(
-                  onPressed: () {
-                    setState(() {
-                      isRegisterPage = false;
-                    });
-                  },
-                  child: LenraText(
-                    text: "Log in",
-                    style: !isRegisterPage
-                        ? TextStyle(color: LenraColorThemeData.lenraWhite)
-                        : TextStyle(color: LenraColorThemeData.lenraBlue),
-                  ),
-                  style: ButtonStyle(
-                    backgroundColor: MaterialStateProperty.resolveWith((states) {
-                      if (!isRegisterPage) {
-                        return LenraColorThemeData.lenraBlue;
-                      }
-                      return LenraColorThemeData.greySuperLight;
-                    }),
-                    shape: MaterialStateProperty.all(
-                      RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(32),
-                      ),
+                LenraFlex(
+                  direction: Axis.vertical,
+                  crossAxisAlignment: CrossAxisAlignment.center,
+                  children: [
+                    LenraText(
+                      text: snapshot.data?.name ?? "App",
+                      style: themeData.lenraTextThemeData.headline1,
                     ),
-                    fixedSize: MaterialStateProperty.all(Size(136, 36)),
-                  ),
-                ),
+                    LenraFlex(
+                      spacing: 8,
+                      crossAxisAlignment: CrossAxisAlignment.center,
+                      children: [
+                        LenraText(text: "by"),
+                        Image.asset(
+                          "assets/images/logo-horizontal-black.png",
+                          scale: 1.25,
+                        ),
+                        IconButton(
+                          onPressed: () {
+                            html.window.open('https://lenra.io', "_blank");
+                          },
+                          icon: Icon(Icons.info_outline),
+                        ),
+                      ],
+                    ),
+                  ],
+                )
               ],
-            ),
-          ),
-          isRegisterPage ? registerForm() : loginForm(),
-        ],
-      ),
-    );
-  }
-
-  Widget registerForm() {
-    return Form(
-      key: _formKey,
-      child: LenraFlex(
-        spacing: 24,
-        direction: Axis.vertical,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          //------Email------
-          LenraTextFormField(
-            label: 'Your email',
-            hintText: 'Email',
-            onChanged: (String value) {
-              email = value;
-            },
-            onSubmitted: (_) {
-              submit();
-            },
-            size: LenraComponentSize.large,
-            type: LenraTextFormFieldType.email,
-            autofillHints: [AutofillHints.email],
-          ),
-          //------Password------
-          LenraTextFormField(
-            label: 'Your password',
-            description: "8 characters, 1 uppercase, 1 lowercase and 1 special character.",
-            obscure: _hidePassword,
-            hintText: 'Password',
-            onSubmitted: (_) {
-              submit();
-            },
-            onChanged: (String value) {
-              password = value;
-            },
-            size: LenraComponentSize.large,
-            type: LenraTextFormFieldType.password,
-            autofillHints: [AutofillHints.newPassword],
-            onSuffixPressed: () {
-              setState(() {
-                _hidePassword = !_hidePassword;
-              });
-            },
-          ),
-          LoadingButton(
-            text: "Create my account",
-            onPressed: () {
-              submit();
-            },
-            loading: isLogging,
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget loginForm() {
-    return Form(
-      key: _formKey,
-      child: LenraFlex(
-        spacing: 24,
-        direction: Axis.vertical,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          LenraTextFormField(
-            label: 'Your email',
-            hintText: 'Email',
-            onChanged: (String value) {
-              email = value;
-            },
-            onSubmitted: (_) {
-              submit();
-            },
-            type: LenraTextFormFieldType.email,
-            size: LenraComponentSize.large,
-            autofillHints: [AutofillHints.email],
-          ),
-          //------Password------
-          LenraTextFormField(
-            label: 'Your password',
-            obscure: _hidePassword,
-            hintText: 'Password',
-            onChanged: (String value) {
-              password = value;
-            },
-            onSubmitted: (_) {
-              submit();
-            },
-            type: LenraTextFormFieldType.password,
-            size: LenraComponentSize.large,
-            autofillHints: [AutofillHints.password],
-            onSuffixPressed: () {
-              setState(() {
-                _hidePassword = !_hidePassword;
-              });
-            },
-          ),
-          LenraFlex(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              LenraCheckbox(
-                value: keep,
-                onPressed: (value) {
-                  setState(() {
-                    keep = value!;
-                  });
-                },
-              ),
-              LenraText(
-                text: "Keep me logged in",
-                style: themeData.lenraTextThemeData.disabledBodyText,
-              ),
-            ],
-          ),
-          LenraFlex(
-            direction: Axis.vertical,
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              LoadingButton(
-                text: "Log in",
-                onPressed: () {
-                  submit();
-                },
-                loading: isLogging,
-              ),
-              Container(
-                padding: EdgeInsets.only(top: 10.0),
-                child: RichText(
-                  text: TextSpan(
-                    text: "I forgot my password",
-                    style: themeData.lenraTextThemeData.blueBodyText,
-                    recognizer: TapGestureRecognizer()
-                      ..onTap = () {
-                        CommonNavigator.go(context, CommonNavigator.lostPassword);
-                      },
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-
-  void submit() {
-    if (_formKey.currentState!.validate()) {
-      AuthModel authModel = context.read<AuthModel>();
-      if (isRegisterPage) {
-        authModel.register(email, password).then((_) {
-          GoRouter.of(context).go(context.read<AuthModel>().redirectToRoute ?? '/');
-        }).catchError((error) {
-          logger.warning(error);
-        });
-      } else {
-        authModel.login(email, password, keep).then((_) {
-          GoRouter.of(context).go(context.read<AuthModel>().redirectToRoute ?? '/');
-        }).catchError((error) {
-          logger.warning(error);
-        });
-      }
+            );
+          }
+        },
+      );
     }
+    return Container();
   }
 }
